Code cleanup/integration & parsing of arbitrary CSV files
Finish adding all of Mia/Kristin's code to the working project structure/design.

Finished the parsing of CSV files from a user specified CSV file.

the command line usage for the driver program is the following, where {p-value} is an optional value:
           driver inputFileName {p-value}
 
the driver exe currently needs to be in the same directory as the input file.